-  Update name of package, and files to match, and add a module.export for inclusion in other scripts, as well as a start.js to auto-start the service -  Bump to 1.0 -  Add in Steve to author list -  ignore default data directory -  update main script to be ./lib/caching-proxy.js -  when doing an npm install of this package, you can use "node start.js" to automatically fire up a standalone cache server -  if require("caching-proxy"), you can start it up by 1) standalone require("caching-proxy").start() or 2) get a reverence to the base CachingProxy via var proxy = require("caching-proxy").CachingProxy, and then new proxy({port: 1000}) -  add a header indicating if served from cache or from a fresh request (headers['served-by'] = "CachingProxy Fresh"; or headers['served-by'] = "CachingProxy Cache";) - rename the forever.js script to daemon.sh to accurately relect it is a shell script (not run via node forever.js, but rather ./daemon.sh)
