+++ lib/cache.js
@@ -178,7 +178,10 @@ Cache.prototype = {
             //console.log('ifModifiedSince', ifModifiedSince, ifModified, lastModified, lastModified > ifModified);
             if(lastModified <= ifModified){
                 console.log('(send304) Serving 304, ' + lastModified + '<=' + ifModified);
-                res.writeHead(304);
+                res.writeHead(304, {
+                    'served-by' : 'CachingProxy Cache'
+                });
+
                 res.end();
                 return true;
             }
@@ -208,6 +211,8 @@ Cache.prototype = {
     serve: function(req, res){
         var metaData = this.getHeadersFromFile(),
             headers = metaData && metaData.headers;
+
+        headers['served-by'] = "CachingProxy Cache";
 
         if(!metaData){
             this.sendHeaders(res, 500, {});
@@ -323,6 +328,9 @@ Cache.prototype = {
         var success = incomingMessage.statusCode === 200;
         var headers = incomingMessage.headers;
 
+        headers['served-by'] = "CachingProxy Fresh";
+
+
         var contentType = headers && headers['content-type'];
         var contentLength = headers && headers['content-length'];
 
@@ -337,14 +345,14 @@ Cache.prototype = {
             incomingMessage.on('end', this.onEndSecure.bind(this, res, incomingMessage));
             incomingMessage.on('data', this.appendData.bind(this));
         }else{
-            this.sendHeaders(res, incomingMessage.statusCode, incomingMessage.headers);
+            this.sendHeaders(res, incomingMessage.statusCode, headers);
             incomingMessage.pipe(res);
 
         }
 
         //pipe to the file system as well
         if(success){
-            this.writeHeaders(incomingMessage.statusCode, incomingMessage.headers);
+            this.writeHeaders(incomingMessage.statusCode, headers);
             this.writeBody(incomingMessage);
         }
 
@@ -390,6 +398,7 @@ Cache.prototype = {
         var length = Buffer.byteLength(body, 'utf8');
 
         headers['content-length'] = length;
+        headers['served-by'] = "CachingProxy Fresh";
 
         this.sendHeaders(res, incomingMessage.statusCode, headers);
