+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "replay-server",
-  "description": "A simple replay server useable in your front-end projects to cache API requests and rewrite their responses as needed to be routed through server",
-  "version": "0.1.0",
-  "author": "Chad Wagner <cwagner@activevideo.com> (http://developer.activevideo.com/)",
+  "name": "caching-proxy",
+  "description": "A caching proxy server useable in your front-end projects to cache API requests and rewrite their responses as needed to be routed through server - for tradeshows, demos (offline and online), data that you know will be retired someday and you want a capture in time that you can reuse, and load testing in shared environments (exmample CloudTV where a server could be running thousands of browser sessions at once and you want to test server scalability independent of APIs an app might depend on, ala activevideo.com)",
+  "version": "1.0.0",
+  "author": "Chad Wagner <cwagner@activevideo.com> (http://developer.activevideo.com/), Steve Oziel <S.Oziel@activevideo.com> ",
   "keywords": [
     "test",
     "testing",
@@ -14,7 +14,7 @@
     "api",
     "server"
   ],
-  "main": "./lib/replay",
+  "main": "lib/caching-proxy",
   "directories": {
     "doc": "./doc",
     "lib": "./lib"
@@ -24,7 +24,7 @@
     "forever-monitor": "^1.5.2",
     "optimist": "",
     "follow-redirects" : "",
-    "mkdirp" : ""
+    "mkdirp" : "",
     "request": "^2.49.0"
   },
   "devDependencies": {},
